yaml examples for a distributed build in a kubernetes cluster
Use the k8s-bldr-api format file: status.yaml to perform an echo test
to validate access to an k8s-bldr exec cluster.

Use the k8s-bldr-api format file: build-hello.yaml to compile a go
hello web app and test connection

The command example uses yaml2json for which a statically linked
binary is provided in the repo for testing.